+++ src/filters/mod.rs
@@ -6,11 +6,14 @@
 pub mod addr;
 pub mod any;
 pub mod body;
+#[cfg(feature = "compression")]
+pub mod compression;
 pub mod cookie;
 pub mod cors;
 pub mod ext;
 pub mod fs;
 pub mod header;
+pub mod host;
 pub mod log;
 pub mod method;
 #[cfg(feature = "multipart")]
@@ -19,6 +22,7 @@ pub mod path;
 pub mod query;
 pub mod reply;
 pub mod sse;
+pub mod trace;
 #[cfg(feature = "websocket")]
 pub mod ws;